unify docs script for single execution
